Added functionality

+++ src/main/scala/Nodes/User.scala
@@ -17,7 +17,7 @@ class User() {
   var email: String = new String()
   var pendingRequests: Map[String, String] = new HashMap[String, String]
   var friends: Map[String, String] = new HashMap[String, String]
-  var posts: HashMap[Int, Post] = new HashMap[Int, Post]()
+  var posts: HashMap[Long, Post] = new HashMap[Long, Post]()
   var pages: Map[String, Page] = new HashMap[String, Page]()
 }
